处理了之前pr的一些冲突问题，拉取上游仓库并完成publish文件夹
