--- app.json
@@ -1,50 +1,0 @@
-{
-  "pages": [
-    "pages/home/home",
-    "pages/logs/logs",
-    "pages/orders/orders",
-    "pages/confession/confession",
-    "pages/personal/personal",
-    "pages/publish/publish",
-    "pages/receipt/receipt",
-    "pages/questionnaire/questionnaire"
-  ],
-  "window": {
-    "backgroundTextStyle": "light",
-    "navigationBarBackgroundColor": "#fff",
-    "navigationBarTitleText": "闲钱汪",
-    "navigationBarTextStyle": "black"
-  },
-  "sitemapLocation": "sitemap.json",
-  "tabBar": {
-    "color": "#a9b7b7",
-    "selectedColor": "#1E90FF",
-    "borderStyle": "white",
-    "list": [
-      {
-        "selectedIconPath": "images/首页-selected.png",
-        "iconPath": "images/首页.png",
-        "pagePath": "pages/home/home",
-        "text": "首页"
-      },
-      {
-        "selectedIconPath": "images/订单-selected.png",
-        "iconPath": "images/订单.png",
-        "pagePath": "pages/orders/orders",
-        "text": "订单"
-      },
-      {
-        "selectedIconPath": "images/表白-selected.png",
-        "iconPath": "images/表白.png",
-        "pagePath": "pages/confession/confession",
-        "text": "表白墙"
-      },
-      {
-        "selectedIconPath": "images/个人-selected.png",
-        "iconPath": "images/个人.png",
-        "pagePath": "pages/personal/personal",
-        "text": "个人"
-      }
-    ]
-  }
-}
