--- app.wxss
@@ -1,10 +1,0 @@
-/**app.wxss**/
-.container {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-  padding: 110rpx 0;
-  box-sizing: border-box;
-}
