+++ code/moneydog/FrontEnd/moneydog/project.config.json
@@ -13,7 +13,7 @@
 	},
 	"compileType": "miniprogram",
 	"libVersion": "2.0.4",
-	"appid": "wx2b05af5f1d0c579c",
+	"appid": "wxf6b84b12c9eedaff",
 	"projectname": "money%20dog",
 	"debugOptions": {
 		"hidedInDevtools": []
