--- pages/confession/confession.json
@@ -1,4 +1,0 @@
-{
-  "component": true,
-  "usingComponents": {}
-}
