--- pages/confession/confession.wxml
@@ -1,2 +1,0 @@
-<!--pages/index3/index3.wxml-->
-<text>表白墙页面</text>
