--- pages/confession/confession.wxss
@@ -1,1 +1,0 @@
-/* pages/index3/index3.wxss */
