--- pages/home/home.js
@@ -1,96 +1,0 @@
-//获取应用实例
-const app = getApp()
-
-Page({
-  data: {
-    userInfo: {},
-    hasUserInfo: false,
-    canIUse: wx.canIUse('button.open-type.getUserInfo'),
-    searchUrl: '/images/search-icon.png',
-    showDialog: false //弹窗
-  },
-  //事件处理函数
-  bindViewTap: function() {
-    wx.navigateTo({
-      url: '../logs/logs'
-    })
-  },
-  //跳转到快递代拿接单界面（可编辑删除）
-  goToReceiptPage: function() {
-    wx.navigateTo({
-      url: '../receipt/receipt',
-    })
-  },
-  //跳转到快递代拿发布界面（可编辑删除）
-  goToPublishPage1: function() {
-    wx.navigateTo({
-      url: '../publish/publish?id=1',//跳转的时候传值，在跳转到的页面的js的page处理id
-    })
-  },
-  goToPublishPage2: function () {
-    wx.navigateTo({
-      url: '../publish/publish?id=2',//跳转的时候传值，在跳转到的页面的js的page处理id
-    })
-  },
-  goToPublishPage3: function () {
-    wx.navigateTo({
-      url: '../publish/publish?id=3',//跳转的时候传值，在跳转到的页面的js的page处理id
-    })
-  },
-  goToPublishPage4: function () {
-    wx.navigateTo({
-      url: '../publish/publish?id=4',//跳转的时候传值，在跳转到的页面的js的page处理id
-    })
-  },
-  goToWenjuanPage: function() {
-    wx.navigateTo({
-      url: '../questionnaire/questionnaire',
-    })
-  },
-  toggleDialog() {
-    this.setData({
-      showDialog: !this.data.showDialog
-    })
-  },
-  onLoad: function () {
-    wx.hideTabBar({})
-    if (app.globalData.userInfo) {
-      this.setData({
-        userInfo: app.globalData.userInfo,
-        hasUserInfo: true
-      })
-      wx.showTabBar({})
-    } else if (this.data.canIUse){
-      // 由于 getUserInfo 是网络请求，可能会在 Page.onLoad 之后才返回
-      // 所以此处加入 callback 以防止这种情况
-      app.userInfoReadyCallback = res => {
-        this.setData({
-          userInfo: res.userInfo,
-          hasUserInfo: true
-        })
-        wx.showTabBar({})
-      }
-    } else {
-      // 在没有 open-type=getUserInfo 版本的兼容处理
-      wx.getUserInfo({
-        success: res => {
-          app.globalData.userInfo = res.userInfo
-          this.setData({
-            userInfo: res.userInfo,
-            hasUserInfo: true
-          })
-          wx.showTabBar({})
-        }
-      })
-    }
-  },
-  getUserInfo: function(e) {
-    console.log(e)
-    app.globalData.userInfo = e.detail.userInfo
-    this.setData({
-      userInfo: e.detail.userInfo,
-      hasUserInfo: true
-    })
-    wx.showTabBar({})
-  }
-})
