--- pages/home/home.json
@@ -1,3 +1,0 @@
-{
-  "usingComponents": {}
-}
