--- pages/home/home.wxss
@@ -1,197 +1,0 @@
-/**index.wxss**/
-.getInfoButton {
-  position: fixed;
-  top: 40%;
-  background-color: #87CEFA;
-}
-
-.userinfo {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.userinfo-avatar {
-  width: 200rpx;
-  height: 200rpx;
-  border-radius: 15%;
-}
-
-.userinfo-nickname {
-  color: #aaa;
-}
-
-.button-area {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-}
-
-.button-row {
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-  justify-content: center;
-}
-
-#wenjuan-area {
-  width: 100%;
-  align-content: center;
-  border: 2px solid black;
-}
-
-.wenjuan-button {
-  position: absolute;
-  top:180rpx;
-  left: 180rpx;
-  z-index: 10;
-  border: 2rpx solid gainsboro;
-  background: white;
-  width: 270rpx;
-  height: 270rpx;
-  border-radius: 270rpx;
-}
-
-.wenjuan-button cover-image {
-  margin-left: 70rpx;
-  margin-top: 30rpx;
-}
-
-.wenjuan-button cover-view {
-  margin-left: 95rpx;
-}
-
-.publish-button {
-  width: 355rpx;
-  height: 160rpx;
-  margin: 10rpx;
-  background: #fff;
-}
-
-.publish-button1 {
-  width: 310rpx;
-  height: 310rpx;
-  background-color: #fff;/* #87CEFA;*/
-  margin: 10rpx;
-  padding-right: 100rpx;
-  padding-bottom: 60rpx;
-}
-.publish-button2 {
-  width: 310rpx;
-  height: 310rpx;
-  background-color: #fff;/* #87CEFA;*/
-  margin: 10rpx;
-  padding-left: 100rpx;
-  padding-bottom: 60rpx;
-}
-.publish-button3 {
-  width: 310rpx;
-  height: 310rpx;
-  background-color: #fff;/* #87CEFA;*/
-  margin: 10rpx;
-  padding-right: 100rpx;
-  padding-top: 60rpx;
-}
-.publish-button4 {
-  width: 310rpx;
-  height: 310rpx;
-  background-color: #fff;/* #87CEFA;*/
-  margin: 10rpx;
-  padding-left: 100rpx;
-  padding-top: 60rpx;
-}
-
-.search {
-  width:600rpx;
-  height: 50rpx;
-  border: 3rpx solid black;
-  position: absolute;
-  top: 1%;
-  padding: 10rpx;
-  display: flex;
-  flex-direction: row;
-}
-
-.search-icon {
-  width: 50rpx;
-  height: 50rpx;
-  position: absolute;
-  right: 20rpx;
-}
-
-.popup-button {
-  width: 300rpx;
-  height: 170rpx;
-  border-top-left-radius: 160rpx;
-  border-top-right-radius: 160rpx;
-  position: absolute;
-  bottom: 0rpx;
-  left: 225rpx;
-  background-color: #87CEFA;
-  color: white;
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-}
-
-.popup-button::after {
-  border: none;
-}
-
-#send-img {
-  width: 70rpx;
-  height: 70rpx;
-  margin-top: 15rpx;
-}
-
-.publish-buttonIcon {
-  width: 90rpx;
-  height: 90rpx;
-  position: absolute;
-  left: 30rpx;
-  top: 10rpx;
-}
-
-.publish-text {
-  position: absolute;
-  left: 170rpx;
-  top: 50rpx;
-}
-
-.receipt-buttonIcon {
-  width: 120rpx;
-  height: 120rpx;
-}
-
-.zan-dialog__mask {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  z-index: 10;
-  background: rgba(0, 0, 0, 0.4); /*设置阴影半透明背景*/
-  display: none;
-}
-
-.zan-dialog__container {
-  position: fixed;
-  bottom: 200rpx;
-  left: 50rpx;
-  width: 660rpx;/*弹窗布局宽*/
-  height: 650rpx;/*弹窗布局高，与下面弹出距离transform有关*/
-  background: #fff;
-  transform: translateY(300%);/*弹框弹出距离，与弹框布局高度有关，如300%表示弹起距离为3倍弹窗高度*/
-  transition: all 0.4s ease;
-  z-index: 12;
-  border-radius: 20rpx;
-  box-shadow: 0px 3px 3px 2px gainsboro;/*弹框的悬浮阴影效果，如不需要可注释该行*/
-}
-
-.zan-dialog--show .zan-dialog__container {
-  transform: translateY(0);
-}
-
-.zan-dialog--show .zan-dialog__mask {
-  display: block;
-}
