--- pages/logs/logs.json
@@ -1,4 +1,0 @@
-{
-  "navigationBarTitleText": "查看启动日志",
-  "usingComponents": {}
-}
