--- pages/orders/orders.json
@@ -1,4 +1,0 @@
-{
-  "component": true,
-  "usingComponents": {}
-}
