--- pages/personal/personal.wxml
@@ -1,2 +1,0 @@
-<!--pages/index4/index4.wxml-->
-<text>个人信息页面</text>
