--- pages/personal/personal.wxss
@@ -1,1 +1,0 @@
-/* pages/index4/index4.wxss */
