--- pages/publish/publish.json
@@ -1,3 +1,0 @@
-{
-  "usingComponents": {}
-}
