--- pages/questionnaire/questionnaire.json
@@ -1,3 +1,0 @@
-{
-  "usingComponents": {}
-}
