--- pages/questionnaire/questionnaire.wxml
@@ -1,84 +1,0 @@
-<!--pages/questionnaire/questionnaire.wxml-->
-<view class='container'>
-  <!-->问卷创建显示<-->
-  <block wx:if="{{!isCreate}}">
-    <view class='wenjuan-welcome'>
-      <label id='welcome-text'>欢迎来到问卷系统！</label>
-      <image id='goutou-img' src="../../images/qiuzhu.png"></image>
-    </view>
-    <view class='wenjuan-info'>
-      <label>问卷名称：</label>
-      <input></input>
-      <label>问卷说明：</label>
-      <input></input>
-      <button type='primary' bindtap='createQuestionnaire'>创建问卷</button>
-    </view>
-  </block>
-  <!-->问卷创建end<-->
-
-  <!-->问卷内容及新建显示<-->
-  <block wx:else>
-    <!-->tab选项<-->
-    <view class='ques-tab'>
-      <label bindtap="toggleDialog">添加题目</label>
-      <label>保存</label>
-    </view>
-    <!-->tab选项end<-->
-
-    <!-->弹窗选择题目类型<-->
-    <view class="zan-dialog {{ showDialog ? 'zan-dialog--show' : '' }}">
-    <view class="zan-dialog__mask" bindtap="toggleDialog" />
-    <view class="zan-dialog__container">
-        <button type='primary' bindtap='createSCQ'>单选题</button>
-        <button type='primary' bindtap='createMCQ'>多选题</button>
-        <button type='primary' bindtap='createCompletion'>填空题</button>
-    </view>
-  </view>
-  <!-->弹窗选择题目类型end<-->
-
-  <!-->单选题创建<-->
-  <view class='choice-question' id="{{isCreateSCQ ? 'showSCQ' : 'noshowSCQ'}}">
-    <view>
-      <label>*标题：</label>
-      <input placeholder="请输入标题"/>
-    </view>
-    <view>
-      <label>*选项：</label>
-      <input placeholder="输入选项"/>
-      <input placeholder="输入选项"/>
-      <input placeholder="输入选项"/>
-      <input placeholder="输入选项"/>
-    </view>
-    <button type='primary' bindtap='confirmSCQ'>确认</button>
-  </view>
-  <!-->单选题创建end<-->
-
-  <!-->多选题创建<-->
-  <view class='choice-question' id="{{isCreateMCQ ? 'showMCQ' : 'noshowMCQ'}}">
-    <view>
-      <label>*标题：</label>
-      <input placeholder="请输入标题"/>
-    </view>
-    <view>
-      <label>*选项：</label>
-      <input placeholder="输入选项"/>
-      <input placeholder="输入选项"/>
-      <input placeholder="输入选项"/>
-      <input placeholder="输入选项"/>
-    </view>
-    <button type='primary' bindtap='confirmMCQ'>确认</button>
-  </view>
-  <!-->多选题创建end<-->
-
-  <!-->填空题创建<-->
-  <view class='choice-question' id="{{isCreateCompletion ? 'showCompletion' : 'noshowCompletion'}}">
-    <view>
-      <label>*标题：</label>
-      <input placeholder="请输入标题"/>
-    </view>
-    <button type='primary' bindtap='confirmCompletion'>确认</button>
-  </view>
-  <!-->填空题创建end<-->
-  </block>
-  <!-->问卷内容及新建显示end<-->
-</view>
