--- pages/questionnaire/questionnaire.wxss
@@ -1,134 +1,0 @@
-/* pages/questionnaire/questionnaire.wxss */
-.wenjuan-welcome {
-  width: 600rpx;
-  height: 150rpx;
-  vertical-align: top;
-  position: absolute;
-  top: 0%;
-  display: flex;
-  flex-direction: row;
-  margin-top: 30rpx;
-}
-
-#welcome-text {
-  padding-right: 80rpx;
-}
-
-#goutou-img {
-  width: 150rpx;
-  height: 150rpx;
-  position: absolute;
-  right: 0rpx;
-}
-
-.wenjuan-info {
-  width: 600rpx;
-  margin-top: 100rpx;
-}
-
-.wenjuan-info input {
-  width: 100%;
-  height: 200rpx;
-  border: 2rpx solid gainsboro;
-  margin-bottom: 40rpx;
-  margin-top: 20rpx;
-}
-
-.wenjuan-info button {
-  margin-top: 60rpx;
-}
-
-.ques-tab {
-  display: flex;
-  flex-direction: row;
-  position: absolute;
-  bottom: 0rpx;
-  height: 100rpx;
-  width: 100%;
-}
-
-.ques-tab label {
-  width: 50%;
-  text-align: center;
-  border: 2rpx solid gainsboro;
-  border-radius: 5%;
-}
-
-.zan-dialog__mask {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  z-index: 10;
-  background: rgba(0, 0, 0, 0.4); /*设置阴影半透明背景*/
-  display: none;
-}
-
-.zan-dialog__container {
-  position: fixed;
-  bottom: 350rpx;
-  left: 100rpx;
-  width: 400rpx;/*弹窗布局宽*/
-  height: 400rpx;/*弹窗布局高，与下面弹出距离transform有关*/
-  background: #fff;
-  transform: translateY(300%);/*弹框弹出距离，与弹框布局高度有关，如300%表示弹起距离为3倍弹窗高度*/
-  transition: all 0.4s ease;
-  z-index: 12;
-  border-radius: 20rpx;
-  box-shadow: 0px 3px 3px 2px gainsboro;/*弹框的悬浮阴影效果，如不需要可注释该行*/
-  padding: 80rpx;
-}
-
-.zan-dialog__container button {
-  width: 400rpx;
-  height: 100rpx;
-  margin-bottom: 60rpx;
-}
-
-.zan-dialog--show .zan-dialog__container {
-  transform: translateY(0);
-}
-
-.zan-dialog--show .zan-dialog__mask {
-  display: block;
-}
-
-.choice-question {
-  background: white;
-  width: 100%;
-  height: 100%;
-}
-
-.choice-question view {
-  padding: 30rpx;
-}
-
-.choice-question view input {
-  border: 2rpx solid gainsboro;
-  margin: 10rpx;
-}
-
-#noshowSCQ {
-  display: none;
-}
-
-#showSCQ {
-  display: block;
-}
-
-#noshowMCQ {
-  display: none;
-}
-
-#showMCQ {
-  display: block;
-}
-
-#noshowCompletion {
-  display: none;
-}
-
-#showCompletion {
-  display: block;
-}
