--- pages/receipt/receipt.json
@@ -1,3 +1,0 @@
-{
-  "usingComponents": {}
-}
