--- pages/receipt/receipt.wxml
@@ -1,2 +1,0 @@
-<!--pages/receipt/receipt.wxml-->
-<text>接单页面</text>
