--- pages/receipt/receipt.wxss
@@ -1,1 +1,0 @@
-/* pages/receipt/receipt.wxss */
